requirements.txt
amqp==5.1.1
billiard==3.6.4.0
blinker==1.5
Bootstrap-Flask==2.0.2
boto3==1.24.40
botocore==1.27.40
celery==5.2.7
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coco-lib==0.1.3
cssmin==0.2.0
dataclasses-json==0.5.7
decorator==5.1.1
dnspython==2.2.1
email-validator==1.2.1
Faker==13.15.1
Flask==2.1.3
Flask-Assets==2.0
Flask-Avatars==0.2.2
Flask-Dropzone==1.6.0
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Moment==1.0.4
Flask-SQLAlchemy==2.5.1
flask-whooshee==0.8.2
Flask-WTF==1.0.1
greenlet==1.1.2
idna==3.3
imageio==2.20.0
importlib-metadata==4.12.0
itsdangerous==2.0.1
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.0
kombu==5.2.4
lxml==4.9.1
MarkupSafe==2.1.1
marshmallow==3.17.0
marshmallow-enum==1.5.1
more-itertools==8.13.0
mypy-extensions==0.4.3
networkx==2.8.5
numpy==1.23.1
opencv-python==4.6.0.66
packaging==21.3
pandas==1.4.3
pathtools==0.1.2
pbr==5.9.0
Pillow==9.2.0
progressbar2==3.55.0
prompt-toolkit==3.0.30
psutil==5.9.1
psycopg2==2.9.3
pyparsing==3.0.9
python-dateutil==2.8.2
python-dotenv==0.20.0
python-utils==3.3.3
pytz==2022.1
PyWavelets==1.3.0
requests==2.28.1
s3transfer==0.6.0
scikit-image==0.19.3
scikit-learn==1.1.1
scikit-video==1.1.11
scipy==1.9.0
six==1.16.0
sklearn==0.0
SQLAlchemy==1.4.39
sqlalchemy-migrate==0.13.0
sqlparse==0.4.2
Tempita==0.5.2
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2022.7.28
typing-inspect==0.7.1
typing_extensions==4.3.0
urllib3==1.26.11
vine==5.0.0
watchdog==2.1.9
wcwidth==0.2.5
webassets==2.0
Werkzeug==2.2.1
Whoosh==2.7.4
WTForms==3.0.1
WTForms-SQLAlchemy==0.3
zipp==3.8.1

# dev
fathomnet==0.5.2
